mh-wasm32: remove unused sharedlib options double-conversion-utils: EMSCRIPTEN is an architecture plurrule: remove an warning on the conversion to double runConfigureICU: more default options nodejs-system: the nodejs system binding

+++ icu4c/source/i18n/plurrule.cpp
@@ -1661,7 +1661,7 @@ int64_t FixedDecimal::getFractionalDigits(double n, int32_t v) {
       case 3: return (int64_t)(fract*1000.0 + 0.5);
       default:
           double scaled = floor(fract * pow(10.0, (double)v) + 0.5);
-          if (scaled > U_INT64_MAX) {
+          if (scaled > (double)U_INT64_MAX) {
               return U_INT64_MAX;
           } else {
               return (int64_t)scaled;
